UHF-6038: Detached fallback navigation from external menu blocks and created a new instance for it.
