+++ helfi_features/helfi_navigation/config/language/fi/block.block.external_footer_top_navigation.yml
@@ -1,0 +1,2 @@
+settings:
+  label: 'Helsingin kaupunki'
