+++ helfi_features/helfi_navigation/config/language/sv/block.block.external_footer_top_navigation_2.yml
@@ -1,0 +1,2 @@
+settings:
+  label: 'Ta kontakt'
