+++ helfi_features/helfi_navigation/config/optional/block.block.external_footer_top_navigation_2.yml
@@ -1,0 +1,22 @@
+langcode: en
+status: true
+dependencies:
+  module:
+    - helfi_navigation
+  theme:
+    - hdbt
+id: external_footer_top_navigation_2
+theme: hdbt
+region: footer_top
+weight: 2
+provider: null
+plugin: 'external_menu_block:footer-top-navigation-2'
+settings:
+  id: 'external_menu_block:footer-top-navigation-2'
+  label: Connect
+  label_display: visible
+  provider: helfi_navigation
+  level: 1
+  depth: 1
+  expand_all_items: false
+visibility: {  }
