+++ helfi_features/helfi_navigation/config/optional/block.block.external_header_top_navigation.yml
@@ -1,0 +1,22 @@
+langcode: en
+status: true
+dependencies:
+  module:
+    - helfi_navigation
+  theme:
+    - hdbt
+id: external_header_top_navigation
+theme: hdbt
+region: header_top
+weight: 0
+provider: null
+plugin: 'external_menu_block:header-top-navigation'
+settings:
+  id: 'external_menu_block:header-top-navigation'
+  label: 'External - Header top navigation'
+  label_display: '0'
+  provider: helfi_navigation
+  level: 1
+  depth: 1
+  expand_all_items: false
+visibility: {  }
