+++ helfi_features/helfi_navigation/src/Plugin/Block/ExternalMenuBlock.php
@@ -5,6 +5,7 @@ declare(strict_types = 1);
 namespace Drupal\helfi_navigation\Plugin\Block;
 
 use Drupal\helfi_api_base\Environment\Project;
+use Drupal\helfi_navigation\ExternalMenuTree;
 
 /**
  * Provides an external menu block.
@@ -19,16 +20,33 @@ use Drupal\helfi_api_base\Environment\Project;
 class ExternalMenuBlock extends ExternalMenuBlockBase {
 
   /**
-   * {@inheritdoc}
+   * Build either fallback menu or external menu tree render array.
+   *
+   * @return array|null
+   *   Returns the render array.
    */
-  public function getData(): string {
+  public function build():? array {
+    $build = [];
+
     $menu_type = $this->getDerivativeId();
 
-    return $this->globalNavigationService->makeRequest(
-      Project::ETUSIVU,
-      'GET',
-      "/global-menus/$menu_type"
+    /** @var \Drupal\helfi_navigation\ExternalMenuTree $menu_tree */
+    $menu_tree = $this->buildFromJson(
+      $this->globalNavigationService->makeRequest(
+        Project::ETUSIVU,
+        'GET',
+        "/global-menus/$menu_type"
+      )
     );
+
+    if ($menu_tree instanceof ExternalMenuTree) {
+      $build['#sorted'] = TRUE;
+      $build['#items'] = $menu_tree->getTree();
+      $build['#theme'] = 'menu__external_menu';
+      $build['#menu_type'] = $menu_type;
+    }
+
+    return $build;
   }
 
 }
