+++ helfi_features/helfi_navigation/templates/menu--external-menu.html.twig
@@ -4,7 +4,7 @@
  * Theme override to display a menu.
  *
  * Available variables:
- * - menu_name: The machine name of the menu.
+ * - menu_type: The machine name of the menu.
  * - items: A nested list of menu items. Each menu item contains:
  *   - attributes: HTML attributes for the menu item.
  *   - below: The menu item child items.
